backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
attrs==25.4.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.14.3
black==25.11.0
boolean.py==5.0
boto3==1.40.76
botocore==1.40.76
build==1.3.0
CacheControl==0.14.4
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.3.1
coloredlogs==15.0.1
cryptography==46.0.3
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
email-validator==2.3.0
emergentintegrations==0.1.0
et_xmlfile==2.0.0
fastapi==0.127.1
fastuuid==0.14.0
filelock==3.20.1
flake8==7.3.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.2.1
google-cloud-videointelligence==2.17.0
google-cloud-vision==3.11.0
google-genai==1.52.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ImageHash==4.3.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jq==1.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==33.1.0
license-expression==30.4.4
limits==5.6.0
litellm==1.80.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
motor==3.7.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
networkx==3.6
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.99.9
openpyxl==3.1.5
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
orjson==3.11.5
overrides==7.7.0
packageurl-python==0.17.6
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pillow==12.0.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.5.0
playwright==1.57.0
pluggy==1.6.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
py-serializable==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.5
pyotp==2.9.0
pyparsing==3.2.5
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==9.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
python-pptx==1.0.2
pytokens==0.3.0
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
s3transfer==0.14.0
s5cmd==0.2.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
starlette==0.49.1
stripe==14.0.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
twilio==9.8.8
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.0
uvicorn==0.25.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.1
xlsxwriter==3.2.9
yarl==1.22.0
zipp==3.23.0
